+++ CHANGELOG.md
@@ -53,6 +53,7 @@
 - Upgraded Josdejong/jsoneditor to 5.5.6
 - Upgraded Juicy/imported-template to 1.4.3
 - Upgraded puppetjs/puppet-polymer-client to 3.1.0 which supports reconnection
+- Changed how Starcounter VS Extension handles JSON files to allow adding existing files without code-behind so that they are not treated as TypedJSON per default [#3075](https://github.com/Starcounter/Starcounter/issues/3075)
 
 ## [2.2.1834] - 2016-04-19
 ### Added